Add methods to work with PayFact,correct some info

+++ src/main/java/com/taxah/diplomdb/repository/PayFactRepository.java
@@ -3,5 +3,5 @@ package com.taxah.diplomdb.repository;
 import com.taxah.diplomdb.model.PayFact;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PayFactRepository extends JpaRepository<PayFact, Integer> {
+public interface PayFactRepository extends JpaRepository<PayFact, Long> {
 }
